+++ tests/framework/grid/GridViewTest.php
@@ -5,7 +5,7 @@ namespace yiiunit\framework\grid;
 
 use yii\data\ArrayDataProvider;
 use yii\grid\DataColumn;
-use yii\grid\GridView;
+use yiiunit\data\grid\TestGridView;
 
 /**
  * @author Evgeniy Tkachenko <et.coder@gmail.com>
@@ -13,12 +13,55 @@ use yii\grid\GridView;
  */
 class GridViewTest extends \yiiunit\TestCase
 {
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->mockApplication();
+    }
+
+    /**
+     * @return array
+     */
+    public function emptyDataProvider()
+    {
+        return [
+            [null, 'No results found.'],
+            ['Empty', 'Empty'],
+            // https://github.com/yiisoft/yii2/issues/13352
+            [false, ''],
+        ];
+    }
+
+    /**
+     * @dataProvider emptyDataProvider
+     * @param mixed $emptyText
+     * @param string $expectedText
+     * @throws \Exception
+     */
+    public function testEmpty($emptyText, $expectedText)
+    {
+        $html = TestGridView::widget([
+            'dataProvider' => new ArrayDataProvider(['allModels' => []]),
+            'showHeader' => false,
+            'emptyText' => $emptyText,
+        ]);
+        $html = preg_replace("/\r|\n/", '', $html);
+
+        if ($expectedText) {
+            $emptyRowHtml = "<tr><td colspan=\"0\"><div class=\"empty\">{$expectedText}</div></td></tr>";
+        } else {
+            $emptyRowHtml = '';
+        }
+        $expectedHtml = "<div><table><tbody>{$emptyRowHtml}</tbody></table></div>";
+
+        $this->assertEquals($expectedHtml, $html);
+    }
+
     public function testGuessColumns()
     {
-        $this->mockApplication();
         $row = ['id' => 1, 'name' => 'Name1', 'value' => 'Value1', 'description' => 'Description1',];
 
-        $grid = new GridView([
+        $grid = new TestGridView([
             'dataProvider' => new ArrayDataProvider(
                 [
                     'allModels' => [
@@ -39,7 +82,7 @@ class GridViewTest extends \yiiunit\TestCase
         $row = array_merge($row, ['relation' => ['id' => 1, 'name' => 'RelationName',],]);
         $row = array_merge($row, ['otherRelation' => (object)$row['relation']]);
 
-        $grid = new GridView([
+        $grid = new TestGridView([
             'dataProvider' => new ArrayDataProvider(
                 [
                     'allModels' => [